+++ packages/ocean-react/src/TransactionListExpandable/index.ts
@@ -1,0 +1,2 @@
+export { default } from './TransactionListExpandable';
+export * from './TransactionListExpandable';