+++ packages/ocean-react/src/_shared/components/AmountDetails/index.ts
@@ -1,0 +1,2 @@
+export { default } from './AmountDetails';
+export * from './AmountDetails';